requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.8
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.0
django-heroku==0.0.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
psycopg2-binary==2.9.2
pytz==2021.3
razorpay==1.2.0
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
